Move all sources related files into own folder
