+++ src/main/apps/dataMigration/legacyMetaFiles.ts
@@ -15,8 +15,11 @@ import { Source } from '../../../ipc/sources';
 import { getAppsRootDir, getNodeModulesDir } from '../../config';
 import { readFile, readJsonFile } from '../../fileUtil';
 import { installedAppPath, writeAppInfo } from '../app';
-import { sourceJsonExistsLocally, writeSourceJson } from '../sourceJson';
-import { writeWithdrawnJson } from '../withdrawnJson';
+import {
+    sourceJsonExistsLocally,
+    writeSourceJson,
+} from '../sources/sourceJson';
+import { writeWithdrawnJson } from '../sources/withdrawnJson';
 
 type AppName = `pc-nrfconnect-${string}`;
