+++ src/main/apps/sources/sourceChanges.ts
@@ -14,11 +14,11 @@ import {
     Source,
     SourceName,
     SourceUrl,
-} from '../../ipc/sources';
-import { getAppsRootDir } from '../config';
-import describeError from '../describeError';
-import { addDownloadAppData } from './app';
-import { downloadAppInfos } from './apps';
+} from '../../../ipc/sources';
+import { getAppsRootDir } from '../../config';
+import describeError from '../../describeError';
+import { addDownloadAppData } from '../app';
+import { downloadAppInfos } from '../apps';
 import { downloadSourceJson, writeSourceJson } from './sourceJson';
 import {
     addToSourceList,
